Margin auto lors du resize des shapes dans une toolbar+changement couleur (rgb 0-255)

+++ src/sample/Model/PolygonJavaFX.java
@@ -18,8 +18,8 @@ public class PolygonJavaFX extends PolygonModel {
         double green = rgb.getG();
         double blue = rgb.getB();
 
-        Double[] rgbTab = new Double[3];
-        rgbTab[0] = red; rgbTab[1] = green; rgbTab[2] = blue;
+        int[] rgbTab = new int[3];
+        rgbTab[0] = (int) red; rgbTab[1] = (int) green; rgbTab[2] = (int) blue;
 
         Double[] points = this.getPoints(this.getNbEdges(), this.getLength());
